REBASE lolwat, INFINITY is a float, not a double
That's POSIX, though:

"INFINITY
A constant expression of type float representing positive or unsigned infinity, if available; else a positive constant of type float that overflows at translation time."
http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/math.h.html

+++ util/parsenum.h
@@ -48,7 +48,7 @@ _Pragma("clang diagnostic pop")
 		errno = 0,						\
 		(((*(x)) = 1, (*(x)) /= 2) > 0)	?			\
 			((*(x)) = parsenum_float((s),			\
-			    -INFINITY, INFINITY)) :			\
+			    (double)-INFINITY, (double)INFINITY)) :	\
 		(((*(x)) = -1) > 0) ?					\
 			((*(x)) = parsenum_unsigned((s), 0, (*(x)))) :	\
 			(assert(0 && "PARSENUM applied to signed integer without specified bounds"), 1),	\
